Move jy2Tsr to utils
