+++ healvis/observatory.py
@@ -17,7 +17,8 @@ import time
 import copy
 
 from .beam_model import PowerBeam, AnalyticBeam
-from .cosmology import jy2Tsr
+from .utils import jy2Tsr
+from .cosmology import c_ms
 
 # Line profiling
 # from line_profiler import LineProfiler
@@ -28,9 +29,6 @@ from .cosmology import jy2Tsr
 # ofile = open('time_profiling.out', 'w')
 # atexit.register(ofile.close)
 # atexit.register(prof.print_stats, stream=ofile)
-
-
-c_ms = c.to('m/s').value
 
 
 @jit
